W3EnvAddressBook, support for Infura and others
New address book implementation that reads the PKs from environment
variables, intended to be used in production.

Also change to support sign-and-send instead of transact
(eth_sendTransaction) because that's how it should be for production
cases using infura or other non-custodial proxies.

Dropped support for Python<3.9

Changes in the tests, to test it (manually) on mumbai testnet.